Add dynamic discount factor in DecisionMaker
Decision making for a case like FTC Autonomous benefits from having
a non-1 discount factor because there is a time limit, but how urgent
things need to be is not constant -- when the end of the period
approaches, urgency increases and tasks that execute after a while may
not be able to finish, so the discount factor increases.

The dynamic discount factor concept implements this change of urgency.
By running a timer within DecisionMaker, it can keep track of time and
how urgent execution needs to be.